+++ src/math/src/params.cpp
@@ -7,7 +7,7 @@ namespace ntru {
 	int G_D = 0;
 	int G_NORM_BOUND = 0;
 	int G_ALPHA = 0;
-	int G_SIGMA = 0;
+	long double G_SIGMA = 0;
 	int G_MAX_SIGN_ATT = 1000;
 
 	double G_NU = 0.0;
